rgw: move rados-specific parts of RGWRealmWatcher into ConfigStore
split RGWRealmWatcher in two halves, moving the rados-specific stuff
into a RadosRealmWatcher that inherits from RGWRealmWatcher

add factory function ConfigStore::create_realm_watcher() that returns a
generic RGWRealmWatcher. ConfigStore backends that don't support
watch/notify can return nullptr

rgw::AppMain uses this to avoid relying on checking for a "rados" driver
name, which doesn't work correctly when there are filters on top of the
RadosStore
