+++ src/rgw/driver/immutable_config/store.h
@@ -15,6 +15,7 @@
 #pragma once
 
 #include "rgw_sal_config.h"
+#include "rgw_zone.h"
 
 namespace rgw::sal {
 
@@ -59,6 +60,10 @@ class ImmutableConfigStore : public ConfigStore {
   virtual int realm_notify_new_period(const DoutPrefixProvider* dpp,
                                       optional_yield y,
                                       const RGWPeriod& period) override;
+  virtual auto create_realm_watcher(const DoutPrefixProvider* dpp,
+                                    optional_yield y,
+                                    const RGWRealm& realm)
+      -> std::unique_ptr<RGWRealmWatcher> override;
   virtual int list_realm_names(const DoutPrefixProvider* dpp,
                                optional_yield y, const std::string& marker,
                                std::span<std::string> entries,
