+++ src/rgw/driver/rados/config/store.h
@@ -66,6 +66,10 @@ class RadosConfigStore : public sal::ConfigStore {
   virtual int realm_notify_new_period(const DoutPrefixProvider* dpp,
                                       optional_yield y,
                                       const RGWPeriod& period) override;
+  virtual auto create_realm_watcher(const DoutPrefixProvider* dpp,
+                                    optional_yield y,
+                                    const RGWRealm& realm)
+      -> std::unique_ptr<RGWRealmWatcher> override;
   virtual int list_realm_names(const DoutPrefixProvider* dpp,
                                optional_yield y, const std::string& marker,
                                std::span<std::string> entries,
